+++ POS/POS/Modules/Payout/PayoutModule.cs
@@ -1,4 +1,5 @@
 ﻿using Framework.Infrastructure.Configuration;
+using Framework.Infrastructure.Data.Configuration;
 using Framework.WPF.ErrorHandling;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -6,6 +7,7 @@ using POS.Common;
 using POS.Core.Config;
 using POS.Infrastructure.Config;
 using POS.Modules.Payout.Services.ViewModels;
+using POS.Modules.Payout.Settings;
 
 namespace POS.Modules.Payout
 {
@@ -13,6 +15,14 @@ namespace POS.Modules.Payout
     {
         public static void AddPayoutModule(this IServiceCollection services, IConfiguration configuration)
         {
+            var supervisorConfigSettings = new SupervisorConfigSettings();
+            services.AddSingleton<IDataConfig>(_ => supervisorConfigSettings);
+            services.AddSingleton(_ => supervisorConfigSettings);
+
+            var cashdrawerConfigSettings = new CashdrawerConfigSettings();
+            services.AddSingleton<IDataConfig>(_ => cashdrawerConfigSettings);
+            services.AddSingleton(_ => cashdrawerConfigSettings);
+
             services.AddTransient<IErrorHandlingService, ErrorHandlerService>();
             services.AddTransient<IMessageBoxService, MessageBoxService>();
             services.AddTransient<IPayoutViewServices, PayoutViewServices>();